+++ CHANGELOG.md
@@ -17,6 +17,16 @@ This project follows **Semantic Versioning (SemVer)** ([semver.org](https://semv
 
 ---
 
+## [0.97.6] - 2026-03-05
+
+### Added
+
+- Add conventional commits regex validation to support squash merge commit messages.
+
+### Update
+
+- bump dorny/test-reporter from 2.5.0 to 2.6.0
+
 ## [0.97.5] - 2026-03-03
 
 ### Added